feat: allow deploying a contract multiple times
