+++ ops/create.ts
@@ -60,7 +60,7 @@ const isValidAddress = (address: string) => {
   }
 }
 
-const isValidAddressOrFail = (address: string) => {
+export const isValidAddressOrFail = (address: string) => {
   if (!isValidAddress(address)) {
     process.exit(1)
   }
@@ -244,8 +244,8 @@ const populateEntries = async (
   return results
 }
 
-const getTokenLockManagerOrFail = async (hre: HardhatRuntimeEnvironment) => {
-  const deployment = await hre.deployments.get('GraphTokenLockManager')
+export const getTokenLockManagerOrFail = async (hre: HardhatRuntimeEnvironment, name: string) => {
+  const deployment = await hre.deployments.get(name)
   if (!deployment.address) {
     logger.error('GraphTokenLockManager address not found')
     process.exit(1)
@@ -275,6 +275,7 @@ task('create-token-locks', 'Create token lock contracts from file')
   .addParam('deployFile', 'File from where to read the deploy config')
   .addParam('resultFile', 'File where to save results')
   .addParam('ownerAddress', 'Owner address of token lock contracts')
+  .addParam('managerName', 'Name of the token lock manager deployment', 'GraphTokenLockManager')
   .addFlag('dryRun', 'Get the deterministic contract addresses but do not deploy')
   .addFlag(
     'txBuilder',
@@ -283,7 +284,7 @@ task('create-token-locks', 'Create token lock contracts from file')
   .addOptionalParam('txBuilderTemplate', 'File to use as a template for the transaction builder')
   .setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
     // Get contracts
-    const manager = await getTokenLockManagerOrFail(hre)
+    const manager = await getTokenLockManagerOrFail(hre, taskArgs.managerName)
 
     // Prepare
     logger.log(await prettyEnv(hre))
@@ -539,132 +540,12 @@ task('create-token-locks-simple', 'Create token lock contracts from file')
     }
   })
 
-task('manager-setup-auth', 'Setup default authorized functions in the manager')
-  .addParam('targetAddress', 'Target address for function calls')
-  .setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
-    // Get contracts
-    const manager = await getTokenLockManagerOrFail(hre)
-
-    // Prepare
-    logger.log(await prettyEnv(hre))
-
-    // Validations
-    isValidAddressOrFail(taskArgs.targetAddress)
-
-    // Setup authorized functions
-    const signatures = [
-      'stake(uint256)',
-      'unstake(uint256)',
-      'withdraw()',
-      'delegate(address,uint256)',
-      'undelegate(address,uint256)',
-      'withdrawDelegated(address,address)',
-      'setDelegationParameters(uint32,uint32,uint32)',
-      'setOperator(address,bool)',
-    ]
-
-    logger.info('The following signatures will be authorized:')
-    logger.info(signatures)
-
-    if (await askConfirm()) {
-      // Setup authorized functions
-      logger.info('Setup authorized functions...')
-      const targets = Array(signatures.length).fill(taskArgs.targetAddress)
-      const tx1 = await manager.setAuthFunctionCallMany(signatures, targets)
-      await waitTransaction(tx1)
-      logger.success('Done!\n')
-
-      // Setup authorized token destinations
-      logger.info('Setup authorized destinations...')
-      const tx2 = await manager.addTokenDestination(taskArgs.targetAddress)
-      await waitTransaction(tx2)
-    }
-  })
-
-task('manager-deposit', 'Deposit fund into the manager')
-  .addParam('amount', 'Amount to deposit in GRT')
-  .setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
-    // Get contracts
-    const manager = await getTokenLockManagerOrFail(hre)
-
-    // Prepare
-    logger.log(await prettyEnv(hre))
-
-    const tokenAddress = await manager.token()
-
-    logger.info('Using:')
-    logger.log(`> GraphToken: ${tokenAddress}`)
-    logger.log(`> GraphTokenLockMasterCopy: ${await manager.masterCopy()}`)
-    logger.log(`> GraphTokenLockManager: ${manager.address}`)
-
-    // Deposit funds
-    logger.log(`You are depositing ${taskArgs.amount} into ${manager.address}...`)
-    if (await askConfirm()) {
-      const weiAmount = parseEther(taskArgs.amount)
-
-      logger.log('Approve...')
-      const grt = await hre.ethers.getContractAt('ERC20', tokenAddress)
-      const tx1 = await grt.approve(manager.address, weiAmount)
-      await waitTransaction(tx1)
-
-      logger.log('Deposit...')
-      const tx2 = await manager.deposit(weiAmount)
-      await waitTransaction(tx2)
-    }
-  })
-
-task('manager-withdraw', 'Withdraw fund from the manager')
-  .addParam('amount', 'Amount to deposit in GRT')
-  .setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
-    // Get contracts
-    const manager = await getTokenLockManagerOrFail(hre)
-
-    // Prepare
-    logger.log(await prettyEnv(hre))
-
-    const tokenAddress = await manager.token()
-
-    logger.info('Using:')
-    logger.log(`> GraphToken: ${tokenAddress}`)
-    logger.log(`> GraphTokenLockMasterCopy: ${await manager.masterCopy()}`)
-    logger.log(`> GraphTokenLockManager: ${manager.address}`)
-
-    // Withdraw funds
-    logger.log(`You are withdrawing ${taskArgs.amount} from ${manager.address}...`)
-    if (await askConfirm()) {
-      const weiAmount = parseEther(taskArgs.amount)
-
-      logger.log('Deposit...')
-      const tx = await manager.withdraw(weiAmount)
-      await waitTransaction(tx)
-    }
-  })
-
-task('manager-balance', 'Get current manager balance').setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
-  // Get contracts
-  const manager = await getTokenLockManagerOrFail(hre)
-
-  // Prepare
-  logger.log(await prettyEnv(hre))
-
-  const tokenAddress = await manager.token()
-  const managerOwnerAddress = await manager.owner()
-
-  logger.info('Using:')
-  logger.log(`> GraphToken: ${tokenAddress}`)
-  logger.log(`> GraphTokenLockMasterCopy: ${await manager.masterCopy()}`)
-  logger.log(`> GraphTokenLockManager: ${manager.address} owner: ${managerOwnerAddress}`)
-
-  const grt = await hre.ethers.getContractAt('ERC20', tokenAddress)
-  const balance = await grt.balanceOf(manager.address)
-  logger.log('Current Manager balance is ', formatEther(balance))
-})
-
 task('scan-token-locks-balances', 'Check current balances of deployed contracts')
   .addParam('resultFile', 'File where to load deployed contracts')
+  .addParam('managerName', 'Name of the token lock manager deployment', 'GraphTokenLockManager')
   .setAction(async (taskArgs, hre: HardhatRuntimeEnvironment) => {
     // Get contracts
-    const manager = await getTokenLockManagerOrFail(hre)
+    const manager = await getTokenLockManagerOrFail(hre, taskArgs.managerName)
 
     // Prepare
     logger.log(await prettyEnv(hre))
